+++ src/prelude.rs
@@ -27,6 +27,6 @@
 //! ```
 
 pub use crate::rgb::{
-    Abgr8888, Rgb888, RgbColor as _, Rgbaf32, Rgbf32, WithBlue as _, WithGreen as _, WithRed as _,
+    Abgr8888, HasBlue as _, HasGreen as _, HasRed as _, Rgb888, RgbColor as _, Rgbaf32, Rgbf32,
 };
 pub use crate::scalar::normalize::{Normalize as _, Normalized as _};